feat(translate): typed keys for translate pipe
